+++ rio_tiler_pds/modis/aws/modis_astraea.py
@@ -5,7 +5,7 @@ from typing import Dict, Type
 import attr
 from morecantile import TileMatrixSet
 
-from rio_tiler.constants import WEB_MERCATOR_TMS
+from rio_tiler.constants import WEB_MERCATOR_TMS, WGS84_CRS
 from rio_tiler.errors import InvalidBandName
 from rio_tiler.io import COGReader, MultiBandReader
 from rio_tiler_pds.errors import InvalidMODISProduct
@@ -100,10 +100,12 @@ class MODISReader(MultiBandReader):
 
     """
 
-    sceneid: str = attr.ib()
+    input: str = attr.ib()
+    tms: TileMatrixSet = attr.ib(default=WEB_MERCATOR_TMS)
+
     reader: Type[COGReader] = attr.ib(default=COGReader)
     reader_options: Dict = attr.ib(factory=dict)
-    tms: TileMatrixSet = attr.ib(default=WEB_MERCATOR_TMS)
+
     minzoom: int = attr.ib(default=4)
     # Most of MODIS product are at 500m resolution (zoom = 8)
     # Some are at 250m (zoom = 10) (MOD09GQ & MYD09GQ) thus we use maxzoom = 9 by default
@@ -115,16 +117,17 @@ class MODISReader(MultiBandReader):
 
     def __attrs_post_init__(self):
         """Parse Sceneid and get grid bounds."""
-        self.scene_params = sceneid_parser(self.sceneid)
+        self.scene_params = sceneid_parser(self.input)
         product = self.scene_params["product"]
         if product not in modis_valid_bands:
             raise InvalidMODISProduct(f"{product} is not supported.")
 
         self.bands = modis_valid_bands[product]
         self.bounds = tile_bbox(
-            self.scene_params["horizontal_grid"], self.scene_params["vertical_grid"],
+            self.scene_params["horizontal_grid"],
+            self.scene_params["vertical_grid"],
         )
-        return self
+        self.crs = WGS84_CRS
 
     def _get_band_url(self, band: str) -> str:
         """Validate band's name and return band's url."""
@@ -141,4 +144,4 @@ class MODISReader(MultiBandReader):
             band_prefix = ""
 
         prefix = self._prefix.format(**self.scene_params)
-        return f"{self._scheme}://{self._hostname}/{prefix}/{self.sceneid}_{band_prefix}{band}.TIF"
+        return f"{self._scheme}://{self._hostname}/{prefix}/{self.input}_{band_prefix}{band}.TIF"